LayoutTests/platform/glib/imported/w3c/web-platform-tests/websockets/constructor/010-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://web-platform.test:49001/handshake_protocol' failed: The server did not accept the WebSocket handshake.

PASS WebSockets: protocol in response but no requested protocol
